Update mutations_game_of_life
(This version is actually from 3/5, but I forgot to upload it then)

Here, I tried to implement what I talked about with Kahini in office hours. I initiated current_gen as a 3-dimensional matrix so that I can manipulate the colors. Then, I concatenated it with itself, and the result was a black and white version of Game of Life, which is what I wanted. Then, I tried to select a random cell and make it red, but I was unsuccessful with this.